+++ Framework/FormHtml.php
@@ -307,6 +307,7 @@ class FormHtml {
         $html .= "</div>";
         $html .= "</div>";
         $html .= "</div>";
+        $html .= "</div>";
         return $html;
     }