+++ Modules/booking/Model/BkCalendarEntry.php
@@ -11,11 +11,15 @@ require_once 'Modules/booking/Model/BkScheduling.php';
 require_once 'Modules/booking/Model/BkNightWE.php';
 
 /**
- * Class defining the GRR area model
+ * Class defining the booking entries
  *
  * @author Sylvain Prigent
  */
 class BkCalendarEntry extends Model {
+
+    public static $REASON_BOOKING = 0;
+    public static $REASON_HOLIDAY = 1;
+    public static $REASON_MAINTENANCE = 2;
 
     public function __construct() {
         $this->tableName = "bk_calendar_entry";
@@ -47,6 +51,7 @@ class BkCalendarEntry extends Model {
         `period_id` int(11) NOT NULL DEFAULT 0,
         `all_day_long` int(1) NOT NULL DEFAULT 0,
         `deleted` int(1) NOT NULL DEFAULT 0,
+        `reason` int NOT NULL DEFAULT 0,
 		PRIMARY KEY (`id`)
 		);";
 
@@ -138,7 +143,6 @@ class BkCalendarEntry extends Model {
         foreach($resps as $resp){
             
             $sqlr = "SELECT name from cl_clients WHERE id=? AND deleted=0 AND id_space=?";
-            //$sqlr = "SELECT name, firstname FROM core_users WHERE id=?";
 
             $respinfo = $this->runRequest($sqlr, array($resp[0], $id_space))->fetch();
             if(!$respinfo) {
@@ -170,7 +174,6 @@ class BkCalendarEntry extends Model {
         $modelUserClient = new ClClientUser();
         
         foreach ($data as $d) {
-            //$resps = $modelUserClient->getUserAccounts($id_space, $d["recipient_id"]);
             $resps = $modelUserClient->getUserClientAccounts($d["recipient_id"], $id_space);
 
             if (!empty($resps)) {
@@ -256,6 +259,7 @@ class BkCalendarEntry extends Model {
             "package_id" => 0,
             "responsible_id" => $resps_id,
             "invoice_id" => 0,
+            "reason" => self::$REASON_BOOKING,
             "all_day_long" => 0);
     }
 
@@ -269,15 +273,15 @@ class BkCalendarEntry extends Model {
         $this->runRequest($sql, array($deleted, $id, $id_space));
     }
 
-    public function setEntry($id_space, $id, $start_time, $end_time, $resource_id, $booked_by_id, $recipient_id, $last_update, $color_type_id, $short_description, $full_description, $quantities, $supplementaries, $package_id, $responsible_id) {
+    public function setEntry($id_space, $id, $start_time, $end_time, $resource_id, $booked_by_id, $recipient_id, $last_update, $color_type_id, $short_description, $full_description, $quantities, $supplementaries, $package_id, $responsible_id, $reason=0) {
         $old = null;
         if (!$id) {
             $sql = "INSERT INTO bk_calendar_entry (start_time, end_time, resource_id, booked_by_id, recipient_id, 
                     last_update, color_type_id, short_description, full_description, quantities, 
-                    supplementaries, package_id, responsible_id, id_space) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?, ?)";
+                    supplementaries, package_id, responsible_id, id_space, reason) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
             $this->runRequest($sql, array($start_time, $end_time, $resource_id, $booked_by_id, $recipient_id,
                 $last_update, $color_type_id, $short_description, $full_description, $quantities,
-                $supplementaries, $package_id, $responsible_id, $id_space));
+                $supplementaries, $package_id, $responsible_id, $id_space, $reason));
             $id = $this->getDatabase()->lastInsertId();
         } else {
             $sql = "SELECT * FROM bk_calendar_entry WHERE id=? AND id_space=?";
@@ -285,10 +289,10 @@ class BkCalendarEntry extends Model {
             $old = ['start_time' => $oldRes['start_time'], 'resource_id' => $oldRes['resource_id'], 'recipient_id' => $oldRes['recipient_id'], 'booked_by_id' => $oldRes['booked_by_id'], 'responsible_id' => $oldRes['responsible_id']];
             $sql = "UPDATE bk_calendar_entry SET start_time=?, end_time=?, resource_id=?, booked_by_id=?, recipient_id=?, 
                     last_update=?, color_type_id=?, short_description=?, full_description=?, quantities=?, 
-                    supplementaries=?, package_id=?, responsible_id=? WHERE id=? AND deleted=0 AND id_space=?";
+                    supplementaries=?, package_id=?, responsible_id=?, reason=? WHERE id=? AND deleted=0 AND id_space=?";
             $this->runRequest($sql, array($start_time, $end_time, $resource_id, $booked_by_id, $recipient_id,
                 $last_update, $color_type_id, $short_description, $full_description, $quantities,
-                $supplementaries, $package_id, $responsible_id, $id, $id_space));
+                $supplementaries, $package_id, $responsible_id, $reason, $id, $id_space));
         }
         Events::send(["action" => Events::ACTION_CAL_ENTRY_EDIT, "bk_calendar_entry_old" => $old, "bk_calendar_entry" => ["id" => intval($id), "id_space" => $id_space]]);
 
@@ -298,25 +302,26 @@ class BkCalendarEntry extends Model {
 
     /**
      * Add a calendar entry
-     * @param unknown $start_time
-     * @param unknown $end_time
-     * @param unknown $resource_id
-     * @param unknown $booked_by_id
-     * @param unknown $recipient_id
-     * @param unknown $last_update
-     * @param unknown $color_type_id
-     * @param unknown $short_description
-     * @param unknown $full_description
+     * @param int $start_time
+     * @param int $end_time
+     * @param int $resource_id
+     * @param int $booked_by_id
+     * @param int $recipient_id
+     * @param int $last_update
+     * @param int $color_type_id
+     * @param string $short_description
+     * @param string $full_description
      * @param number $quantity
+     * @param int package
      * @return string
      */
-    public function addEntry($id_space, $start_time, $end_time, $resource_id, $booked_by_id, $recipient_id, $last_update, $color_type_id, $short_description, $full_description, $quantity = 0, $package = 0) {
+    public function addEntry($id_space, $start_time, $end_time, $resource_id, $booked_by_id, $recipient_id, $last_update, $color_type_id, $short_description, $full_description, $quantity = 0, $package = 0, $reason=0) {
 
         $sql = "insert into bk_calendar_entry(start_time, end_time, resource_id, booked_by_id, recipient_id, 
-							last_update, color_type_id, short_description, full_description, quantities, package_id, id_space)"
-                . " values(?,?,?,?,?,?,?,?,?,?,?,?)";
+							last_update, color_type_id, short_description, full_description, quantities, package_id, id_space,supplementaries, reason)"
+                . " values(?,?,?,?,?,?,?,?,?,?,?,?,?,?)";
         $this->runRequest($sql, array($start_time, $end_time, $resource_id, $booked_by_id, $recipient_id,
-            $last_update, $color_type_id, $short_description, $full_description, $quantity, $package, $id_space));
+            $last_update, $color_type_id, $short_description, $full_description, $quantity, $package, $id_space, '', $reason));
         return $this->getDatabase()->lastInsertId();
     }
 
@@ -361,17 +366,19 @@ class BkCalendarEntry extends Model {
 
     /**
      * Add a calendar entry if not exists  
-     * @param unknown $id
-     * @param unknown $start_time
-     * @param unknown $end_time
-     * @param unknown $resource_id
-     * @param unknown $booked_by_id
-     * @param unknown $recipient_id
-     * @param unknown $last_update
-     * @param unknown $color_type_id
-     * @param unknown $short_description
-     * @param unknown $full_description
+     * @param int $id_space
+     * @param int $id
+     * @param int $start_time
+     * @param int $end_time
+     * @param int $resource_id
+     * @param int $booked_by_id
+     * @param int $recipient_id
+     * @param int $last_update
+     * @param int $color_type_id
+     * @param string $short_description
+     * @param string $full_description
      * @param number $quantity
+     * @param int package
      */
     public function setEntryOld($id_space, $id, $start_time, $end_time, $resource_id, $booked_by_id, $recipient_id, $last_update, $color_type_id, $short_description, $full_description, $quantity = 0, $package = 0) {
 
@@ -475,12 +482,9 @@ class BkCalendarEntry extends Model {
         $req = $this->runRequest($sql, $q);
         $data = $req->fetchAll(); // Liste des bénéficiaire dans la période séléctionée
 
-        //$modelUser = new CoreUser();
-        //$modelColor = new BkColorCode();
         for ($i = 0; $i < count($data); $i++) {
             $rid = $data[$i]["recipient_id"];
             if ($rid > 0) {
-                //$userInfo = $modelUser->userAllInfo($rid);
                 $data[$i]["recipient_fullname"] = $data[$i]["lastname"] . " " . $data[$i]["firstname"];
             } else {
                 $data[$i]["recipient_fullname"] = "";
@@ -528,7 +532,6 @@ class BkCalendarEntry extends Model {
         for ($i = 0; $i < count($data); $i++) {
             $rid = $data[$i]["recipient_id"];
             if ($rid > 0) {
-                //$userInfo = $modelUser->userAllInfo($rid);
                 $data[$i]["recipient_fullname"] = $data[$i]["lastname"] . " " . $data[$i]["firstname"];
             } else {
                 $data[$i]["recipient_fullname"] = "";
@@ -547,24 +550,15 @@ class BkCalendarEntry extends Model {
 
     /**
      * Get entries for a given period and a given area
-     * @param unknown $dateBegin
-     * @param unknown $dateEnd
-     * @param unknown $areaId
+     * @param int $dateBegin
+     * @param int $dateEnd
+     * @param int $areaId
      * @return multitype:
      */
     public function getEntriesForPeriodeAndArea($id_space, $dateBegin, $dateEnd, $areaId, $id_user='') {
 
         $modelResource = new ResourceInfo();
         $resources = $modelResource->resourceIDNameForArea($id_space, $areaId);
-
-        /*
-        $data = array();
-        foreach ($resources as $resource) {
-            $id = $resource["id"];
-            $dataInter = $this->getEntriesForPeriodeAndResource($id_space, $dateBegin, $dateEnd, $id);
-            $data = array_merge($data, $dataInter);
-        }
-        */
 
         $rids = [];
         foreach($resources as $r) {
@@ -776,7 +770,6 @@ class BkCalendarEntry extends Model {
             $req = $this->runRequest($sql, array($rec['recipient_id']));
             $resp_id_req = $req->fetch();
             $resp_id_req = $resp_id_req[0];
-            //echo "resp_id_req = " . $resp_id_req . "<br />";
             if ($resp_id_req == $unit_id) {
                 return true;
             }
@@ -822,7 +815,8 @@ class BkCalendarEntry extends Model {
 
     /**
      * Get the emails address of the users who booked a resource and still have a role in space
-     * @param unknown $resource_id
+     * @param int $resource_id
+     * @param int $ts  select reservations after timestamp
      * @return multitype:
      */
     public function getEmailsBookerResource($id_space, $resource_id, $ts=0) {
@@ -845,7 +839,7 @@ class BkCalendarEntry extends Model {
 
     /**
      * Get the emails address of the users who booked resorces of a given area
-     * @param unknown $area_id
+     * @param int $area_id
      * @return multitype:
      */
     public function getEmailsBookerArea($id_space, $area_id) {
@@ -902,6 +896,12 @@ class BkCalendarEntry extends Model {
         $q = array('today' => $now, 'id_user' => $id_user, 'id_space' => $id_space);
         $res = $this->runRequest($sql, $q);
         return $res->fetchAll();
+    }
+
+    public function blockedEntries($id_space) {
+        $sql = "SELECT * FROM bk_calendar_entry WHERE reason>0 AND id_space=? AND deleted=0 ORDER BY start_time DESC";
+        $req = $this->runRequest($sql, array($id_space));
+        return $req->fetchAll();
     }
 
     public function computeDuration($id_space, $booking) {