+++ Modules/booking/Model/BookinginvoiceTranslator.php
@@ -182,7 +182,7 @@ class BookinginvoiceTranslator {
         if ($lang == "fr") {
             return "Bloquer ressources";
         }
-        return "Block resouces";
+        return "Block resources";
     }
 
     public static function Edit_color_code($lang) {