+++ Modules/layout.php
@@ -9,6 +9,9 @@
             if (isset($metadesc)) {echo "<meta name=\"description\" content=\"$metadesc\"/>\n";}
         ?>
         <meta name="mode" description="<?php echo $context['dev'] ? 'dev' : 'prod' ?>">
+        <?php startblock('meta') ?>
+
+        <?php endblock() ?>
         <base href="<?php echo  $context['rootWeb'] ?>" >
         <title>
             <?php startblock('title') ?>
@@ -63,9 +66,7 @@
                         <div class="container">
                             <div class="alert alert-warning alert-dismissible  show" role="alert">
                                 <?php echo $context['maintenance']; ?>
-                                <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                                    <span aria-hidden="true">&times;</span>
-                                </button>
+                                <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
                             </div>
                         </div>                        
                     <?php } ?>